use floats for cooldowns

+++ src/structs.h
@@ -35,7 +35,7 @@ typedef struct{
     Texture *tx;
     int x,y;
     int speed_x, speed_y; //used for gravity and jumping
-    int rl_cooldown;
+    float rl_cooldown;
     bool falling;
     int pickup, pickup_active;
     float slow_fall; //1 means no slow fall
@@ -44,7 +44,7 @@ typedef struct{
     int flip; //1 means befault, -1 means flipped
     Color color;
     int state;
-    int anim_cooldown;
+    float anim_cooldown;
     int frame;
 } Soldier;
